(EAI-261) [UI] Code split UI bundle
