+++ packages/mongodb-chatbot-ui/.gitignore
@@ -24,3 +24,6 @@ build-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# Build analysis
+stats.html
