+++ packages/mongodb-chatbot-ui/package.json
@@ -25,14 +25,16 @@
     "build",
     "README.md"
   ],
-  "main": "./build/mongodb-chatbot-ui.umd.js",
-  "module": "./build/mongodb-chatbot-ui.es.js",
+  "module": "./build/index.js",
   "types": "./build/types/index.d.ts",
   "exports": {
     ".": {
-      "import": "./build/mongodb-chatbot-ui.es.js",
-      "require": "./build/mongodb-chatbot-ui.umd.js",
+      "import": "./build/index.js",
       "types": "./build/types/index.d.ts"
+    },
+    "./*": {
+      "import": "./build/*.js",
+      "types": "./build/types/*.d.ts"
     }
   },
   "scripts": {
@@ -48,6 +50,7 @@
     "format:check": "prettier ./src --check",
     "lint": "eslint ./src --ext ts,tsx,js,jsx --report-unused-disable-directives",
     "lint:fix": "npm run lint -- --fix && prettier ./src --check --write",
+    "package": "rm -rf build && npm run build:component && npm pack",
     "preview": "vite preview --config vite.config.component.js --mode development",
     "preview:static-site:dev": "npm run build:static-site:dev && vite preview --base ./ --config vite.config.static-site.js",
     "preview:static-site:staging": "npm run build:static-site:staging && vite preview --base ./ --config vite.config.static-site.js",
@@ -64,34 +67,34 @@
   ],
   "dependencies": {
     "@emotion/css": "^11.11.2",
-    "@leafygreen-ui/badge": "^8.0.11",
-    "@leafygreen-ui/banner": "^7.0.12",
-    "@leafygreen-ui/box": "^3.1.7",
-    "@leafygreen-ui/button": "^21.0.3",
-    "@leafygreen-ui/card": "^10.0.3",
-    "@leafygreen-ui/code": "^14.2.12",
-    "@leafygreen-ui/icon": "^11.22.1",
-    "@leafygreen-ui/icon-button": "^15.0.16",
-    "@leafygreen-ui/leafygreen-provider": "^3.1.6",
-    "@leafygreen-ui/loading-indicator": "^2.0.3",
-    "@leafygreen-ui/menu": "^22.0.5",
-    "@leafygreen-ui/modal": "^16.0.3",
-    "@leafygreen-ui/palette": "^4.0.7",
-    "@leafygreen-ui/skeleton-loader": "^1.0.5",
-    "@leafygreen-ui/text-input": "^12.1.18",
-    "@leafygreen-ui/toggle": "^10.0.15",
-    "@leafygreen-ui/typography": "^16.5.4",
-    "@lg-chat/avatar": "^3.0.0",
-    "@lg-chat/chat-disclaimer": "^3.0.0",
-    "@lg-chat/chat-window": "^2.0.0",
-    "@lg-chat/fixed-chat-window": "^2.0.0",
-    "@lg-chat/input-bar": "^4.0.0",
+    "@leafygreen-ui/badge": "^8.1.2",
+    "@leafygreen-ui/banner": "^7.0.20",
+    "@leafygreen-ui/box": "^3.1.9",
+    "@leafygreen-ui/button": "^21.2.0",
+    "@leafygreen-ui/card": "^10.0.7",
+    "@leafygreen-ui/code": "^14.3.2",
+    "@leafygreen-ui/icon": "^11.29.1",
+    "@leafygreen-ui/icon-button": "^15.0.21",
+    "@leafygreen-ui/leafygreen-provider": "^3.1.12",
+    "@leafygreen-ui/loading-indicator": "^2.0.12",
+    "@leafygreen-ui/menu": "^22.0.9",
+    "@leafygreen-ui/modal": "^16.0.7",
+    "@leafygreen-ui/palette": "^4.0.10",
+    "@leafygreen-ui/skeleton-loader": "^1.2.0",
+    "@leafygreen-ui/text-input": "^12.1.27",
+    "@leafygreen-ui/toggle": "^10.1.1",
+    "@leafygreen-ui/typography": "^16.5.5",
+    "@lg-chat/avatar": "^3.0.1",
+    "@lg-chat/chat-disclaimer": "^3.0.2",
+    "@lg-chat/chat-window": "^2.0.1",
+    "@lg-chat/fixed-chat-window": "^2.0.2",
+    "@lg-chat/input-bar": "^4.0.3",
     "@lg-chat/leafygreen-chat-provider": "^2.0.0",
-    "@lg-chat/message": "^4.1.0",
-    "@lg-chat/message-feed": "^3.0.0",
-    "@lg-chat/message-feedback": "^2.0.0",
-    "@lg-chat/message-prompts": "^2.0.0",
-    "@lg-chat/message-rating": "^2.0.0",
+    "@lg-chat/message": "^4.1.1",
+    "@lg-chat/message-feed": "^3.0.1",
+    "@lg-chat/message-feedback": "^2.1.0",
+    "@lg-chat/message-prompts": "^2.0.2",
+    "@lg-chat/message-rating": "^2.0.2",
     "@microsoft/fetch-event-source": "^2.0.1",
     "buffer": "^6.0.3",
     "classnames": "^2.5.1",
@@ -132,6 +135,7 @@
     "react-refresh": "^0.14.0",
     "release-it": "^16.1.5",
     "rollup": "^3.26.3",
+    "rollup-plugin-visualizer": "^5.12.0",
     "storybook": "^7.0.20",
     "typescript": "^5.1.6",
     "vite": "^4.4.7",
