+++ packages/mongodb-chatbot-ui/src/useConversationContext.tsx
@@ -1,7 +1,12 @@
 import { useContext } from "react";
 import { ConversationContext } from "./ConversationProvider";
 
-export default function useConversation() {
+export default function useConversationContext() {
   const conversation = useContext(ConversationContext);
+  if (!conversation) {
+    throw new Error(
+      "useConversationContext must be used within a ChatbotProvider"
+    );
+  }
   return conversation;
 }
